perf: use workerpools to print asset sizes in prod builds
