+++ lib/models/asset-size-printer.js
@@ -2,8 +2,8 @@
 
 const chalk = require('chalk');
 const path = require('path');
-const util = require('util');
 const walkSync = require('walk-sync');
+const workerpool = require('workerpool');
 
 module.exports = class AssetPrinterSize {
   constructor(options) {
@@ -49,11 +49,11 @@ module.exports = class AssetPrinterSize {
 
   makeAssetSizesObject() {
     return new Promise((resolve) => {
-      const fs = require('fs');
-      const zlib = require('zlib');
-      let gzip = util.promisify(zlib.gzip);
       let files = this.findFiles();
       let testFileRegex = /(test-(loader|support))|(testem)/i;
+
+      // create a dedicated worker
+      const pool = workerpool.pool(`${__dirname}/worker.js`);
 
       let assets = files
         // Skip test files
@@ -63,16 +63,16 @@ module.exports = class AssetPrinterSize {
         })
         // Print human-readable file sizes (including gzipped)
         .map((file) => {
-          let contentsBuffer = fs.readFileSync(file);
-          return gzip(contentsBuffer).then((buffer) => ({
-            name: file,
-            size: contentsBuffer.length,
-            gzipSize: buffer.length,
-            showGzipped: contentsBuffer.length > 0,
-          }));
+          return pool
+            .exec('gzipStats', [file])
+            .then((result) => result)
+            .catch((err) => console.error(err));
         });
 
-      return resolve(Promise.all(assets));
+      return Promise.all(assets).then((data) => {
+        pool.terminate(); // terminate all workers when done
+        resolve(data);
+      });
     });
   }
