+++ lib/models/worker.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const workerpool = require('workerpool');
+const gzipStats = require('./gzipStats');
+
+// create worker and register public functions
+workerpool.worker({
+  gzipStats,
+});
